requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.3
click==6.7
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.1
Django==1.9.7
django-filter==1.0.4
djangorestframework==3.6.4
google-api-python-client==1.6.2
gunicorn==19.6.0
httplib2==0.10.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
oauth2client==4.1.1
psycopg2==2.6.2
pyasn1==0.2.3
pyasn1-modules==0.0.9
Pygments==2.2.0
requests==2.17.3
rsa==3.4.2
six==1.10.0
slackclient==1.0.5
uritemplate==3.0.0
urllib3==1.21.1
v==2.1.0
websocket-client==0.40.0
python-dotenv==0.7.1
python-env==1.0.0
